Relocate RegisterMemberCntrllr + ListCoopPresenter
