+++ arbeitszeit_flask/company/routes.py
@@ -91,7 +91,6 @@ from arbeitszeit_web.get_company_transactions import GetCompanyTransactionsPrese
 from arbeitszeit_web.get_plan_summary_company import GetPlanSummaryCompanyPresenter
 from arbeitszeit_web.get_statistics import GetStatisticsPresenter
 from arbeitszeit_web.hide_plan import HidePlanPresenter
-from arbeitszeit_web.list_all_cooperations import ListAllCooperationsPresenter
 from arbeitszeit_web.list_plans import ListPlansPresenter
 from arbeitszeit_web.presenters.company_purchases_presenter import (
     CompanyPurchasesPresenter,
@@ -105,6 +104,9 @@ from arbeitszeit_web.presenters.get_company_summary_presenter import (
 )
 from arbeitszeit_web.presenters.get_coop_summary_presenter import (
     GetCoopSummarySuccessPresenter,
+)
+from arbeitszeit_web.presenters.list_all_cooperations_presenter import (
+    ListAllCooperationsPresenter,
 )
 from arbeitszeit_web.presenters.query_companies_presenter import QueryCompaniesPresenter
 from arbeitszeit_web.presenters.show_a_account_details_presenter import (
