+++ tests/presenters/test_list_all_cooperations.py
@@ -6,7 +6,9 @@ from arbeitszeit.use_cases.list_all_cooperations import (
     ListAllCooperationsResponse,
     ListedCooperation,
 )
-from arbeitszeit_web.list_all_cooperations import ListAllCooperationsPresenter
+from arbeitszeit_web.presenters.list_all_cooperations_presenter import (
+    ListAllCooperationsPresenter,
+)
 from arbeitszeit_web.session import UserRole
 from tests.session import FakeSession
